test: Make getFCOSDownload less arch-dependent
It now supports amd64, arm64, ppc64le and s390x, which are the 4 arches
for which there are fcos images.
